Exract some parts of pubspec parsing to a null-safe library.
